split rc.lua into a modular library «default» and small rc.lua with a bunch of imports
