+++ lib/default/default/bindings/global_mouse.lua
@@ -1,0 +1,14 @@
+local awful = require("awful")
+
+local common = require("default.common")
+local modkey = common.modkey
+
+
+-- {{{ Mouse bindings
+awful.mouse.append_global_mousebindings({
+    awful.button({ }, 3, function () common.mymainmenu:toggle() end),
+    awful.button({ }, 4, awful.tag.viewprev),
+    awful.button({ }, 5, awful.tag.viewnext),
+})
+-- }}}
+
