+++ lib/default/default/focus.lua
@@ -1,0 +1,5 @@
+
+
+client.connect_signal("mouse::enter", function(c)
+    c:activate { context = "mouse_enter", raise = false }
+end)
