+++ lib/default/default/keyboardlayout.lua
@@ -1,0 +1,6 @@
+local awful = require("awful")
+-- Keyboard map indicator and switcher
+
+local mykeyboardlayout = awful.widget.keyboardlayout()
+
+table.insert(globalwidgets.right, mylauncher)
