+++ lib/default/default/layoutbox.lua
@@ -1,0 +1,21 @@
+local awful = require("awful")
+
+    -- Create an imagebox widget which will contain an icon indicating which layout we're using.
+    -- We need one layoutbox per screen.
+local result  = {
+    perscreen = true
+}
+local function worker(s)
+    s.mylayoutbox = awful.widget.layoutbox {
+        screen  = s,
+        buttons = {
+            awful.button({ }, 1, function () awful.layout.inc( 1) end),
+            awful.button({ }, 3, function () awful.layout.inc(-1) end),
+            awful.button({ }, 4, function () awful.layout.inc(-1) end),
+            awful.button({ }, 5, function () awful.layout.inc( 1) end),
+        }
+    }
+    return s.mylayoutbox
+end
+
+table.insert(globalwidgets.right, setmetatable(result, { __call = function(_,...) return worker(...) end}))
