+++ lib/default/default/tags.lua
@@ -1,0 +1,6 @@
+local awful = require("awful")
+
+screen.connect_signal("request::desktop_decoration", function(s)
+    -- Each screen has its own tag table.
+    awful.tag({ "1", "2", "3", "4", "5", "6", "7", "8", "9" }, s, awful.layout.layouts[1])
+end)
