+++ lib/default/default/tasklist.lua
@@ -1,0 +1,25 @@
+local awful = require("awful")
+
+
+
+local result  = {
+    perscreen = true
+}
+local function worker(s)
+    -- Create a tasklist widget
+    s.mytasklist = awful.widget.tasklist {
+        screen  = s,
+        filter  = awful.widget.tasklist.filter.currenttags,
+        buttons = {
+            awful.button({ }, 1, function (c)
+                c:activate { context = "tasklist", action = "toggle_minimization" }
+            end),
+            awful.button({ }, 3, function() awful.menu.client_list { theme = { width = 250 } } end),
+            awful.button({ }, 4, function() awful.client.focus.byidx(-1) end),
+            awful.button({ }, 5, function() awful.client.focus.byidx( 1) end),
+        }
+    }
+    return s.mytasklist
+end
+
+table.insert(globalwidgets.middle, setmetatable(result, { __call = function(_,...) return worker(...) end}))
