+++ lib/default/default/textclock.lua
@@ -1,0 +1,6 @@
+local wibox = require("wibox")
+
+-- Create a textclock widget
+local mytextclock = wibox.widget.textclock()
+
+table.insert(globalwidgets.right, mytextclock)
